+++ NEWS.md
@@ -1,9 +1,12 @@
-# sharpshootR 1.11 (2022-09-21)
+# sharpshootR 1.11 (2022-09-22)
  * adding SoilTaxonomy package to suggests, to allow for better encoding of taxa levels by `SoilTaxonomyDendrogram`
  * `SoilTaxonomyDendrogram()` gains argument `KST.order` to adjust encoding / ordering criteria
  * bug fix in `monthlyWB_summary()` to address warning / `inf` when 0 dry days
  * `plotWB()` aesthetic fixes and additional arguments, including manual specification of y-axis limits
  * `aggregateColorPlot()` gets a new argument for setting label orientation
+ * geomorphic `viz*` functions now use an iterative approach to re-arrange dendrogram according to "hydologic ordering"
+ * new function `plotGeomorphCrossSection()` presents a `SoilProfileCollection` aligned to a geomorphic summary as cross-section
+ * new function `reconcileOSDGeomorp()` convenience function for reconciling SPC and geomorphic summary tables as returned by `soilDB::soilDB()`
 
 # sharpshootR 1.9.1 (2022-02-16)
  * bugfix in `plotWB()` related to figure y-scale range